# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "a full stack Ecommerce app",
    "version": "1.0.0",
    "title": "Ecommerce",
    "contact": {
      "email": "allenyin@cmail.carleton.ca"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": ["http", "https"],
  "paths": {
...
    "/departments": {
      "get": {
        "tags": ["departments"],
        "summary": "get all departments",
        "parameters": [
          {
            "name": "user_token",
            "in": "header",
            "description": "authorization",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "get all departments success",
            "schema": {
              "properties": {
                "departments": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "5cb7cd503f467927c8313c3c"
                      },
                      "departmentName": {
                        "type": "string",
                        "example": "Women"
                      },
                      "categories": {
                        "type": "string",
                        "example": "Basics,Blazer"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "authentication fail",
            "schema": {
              "$ref": "#/definitions/invalid_token"
            }
          },
          "500": {
            "description": "get all departments fail"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": ["categories"],
        "summary": "get all categories",
        "parameters": [
          {
            "name": "user_token",
            "in": "header",
            "description": "authorization",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "get all categories success",
            "schema": {
              "properties": {
                "categories": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "5cb7cd503f467927c8313c38"
                      },
                      "categoryName": {
                        "type": "string",
                        "example": "Blazer"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "authentication fail",
            "schema": {
              "$ref": "#/definitions/invalid_token"
            }
          },
          "500": {
            "description": "get all categories fail"
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": ["products"],
        "summary": "get all products",
        "parameters": [
          {
            "name": "department",
            "in": "query",
            "description": "the department Men and Women...",
            "required": false,
            "type": "string"
          },
          {
            "name": "category",
            "in": "query",
            "description": "the category Jeans,Sweater,etc...",
            "required": false,
            "type": "string"
          },
          {
            "name": "order",
            "in": "query",
            "description": "price, -price...",
            "required": false,
            "type": "string"
          },
          {
            "name": "range",
            "in": "query",
            "description": "price range '20-30'",
            "required": false,
            "type": "string"
          },
          {
            "name": "user_token",
            "in": "header",
            "description": "authorization",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "get all products success",
            "schema": {
              "$ref": "#/definitions/products"
            }
          },
          "401": {
            "description": "authentication fail",
            "schema": {
              "$ref": "#/definitions/invalid_token"
            }
          },
          "404": {
            "description": "get all products fail"
          }
        }
      }
    },
    "/products/{productID}": {
      "get": {
        "tags": ["products"],
        "summary": "get one product data",
        "parameters": [
          {
            "name": "productID",
            "in": "path",
            "description": "The product ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "user_token",
            "in": "header",
            "description": "authorization",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "find product success",
            "schema": {
              "$ref": "#/definitions/product"
            }
          },
          "401": {
            "description": "authentication fail",
            "schema": {
              "$ref": "#/definitions/invalid_token"
            }
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
...
    "/search": {
      "get": {
        "tags": ["search"],
        "summary": "query products by title or category",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "the query Ripped Jeans, Jeans,etc...",
            "required": true,
            "type": "string"
          },
          {
            "name": "order",
            "in": "query",
            "description": "price, -price",
            "required": false,
            "type": "string"
          },
          {
            "name": "range",
            "in": "query",
            "description": "price range '20-30'",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "find products success",
            "schema": {
              "$ref": "#/definitions/products"
            }
          },
          "401": {
            "description": "authentication fail",
            "schema": {
              "$ref": "#/definitions/invalid_token"
            }
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/filter": {
      "get": {
        "tags": ["filter"],
        "summary": "autocomplete input any string return suggestions object",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "find products success",
            "schema": {
              "properties": {
                "filter": {
                  "type": "object",
                  "properties": {
                    "department": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "Men"
                      }
                    },
                    "category": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "Jeans"
                      }
                    },
                    "title": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "Perl Knit Swear"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "authentication fail",
            "schema": {
              "$ref": "#/definitions/invalid_token"
            }
          },
          "404": {
            "description": "not found"
          }
        }
      }
    }
  },
  "definitions": {
...
    "invalid_token": {
      "type": "object",
      "properties": {
        "status": {
          "type": "number",
          "example": 401.0
        },
        "type": {
          "type": "string",
          "example": "invalid_field"
        },
        "error": {
          "$ref": "#/definitions/invalid_token_error"
        }
      }
    },
...
    "productObj": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "5bedf5eec14d7822b39d9d4e"
        },
        "imagePath": {
          "type": "string",
          "example": "/uploads/1775300615_1_1_1.jpg"
        },
        "title": {
          "type": "string",
          "example": "Perl Knit Swear"
        },
        "description": {
          "type": "string",
          "example": "Purl-stitch knit sweater in a combination of textures. Ribbed trim."
        },
        "price": {
          "type": "number",
          "example": 79.99
        },
        "color": {
          "type": "string",
          "example": "Orange"
        },
        "size": {
          "type": "string",
          "example": "M,L"
        },
        "quantity": {
          "type": "number",
          "example": 5.0
        },
        "department": {
          "type": "string",
          "example": "Men"
        },
        "category": {
          "type": "string",
          "example": "Knitwear"
        }
      }
    },
    "product": {
      "type": "object",
      "properties": {
        "product": {
          "$ref": "#/definitions/productObj"
        }
      }
    },
    "products": {
      "type": "object",
      "properties": {
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/productObj"
          }
        }
      }
    },
...
    "invalid_token_error": {
      "properties": {
        "message": {
          "type": "string",
          "example": "Token is not valid"
        }
      }
    },
...
  }
}